% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data_endometrial}
\alias{data_endometrial}
\title{Endometrial Cancer Histology Grade Data}
\format{
A data frame with 79 observations and 4 variables:
\describe{
  \item{NV}{Neovascularization status (0 = absent, 1 = present)}
  \item{PI}{Pulsatility index of the uterine artery}
  \item{EH}{Endometrium height}
  \item{HG}{Histology grade (0 = low grade, 1 = high grade)}
}
}
\source{
\pkg{brglm2} package. The packaged dataset was sourced from
\url{https://users.stat.ufl.edu/~aa/glm/data/}, the data repository used in
Agresti (2015). Originally analyzed in Heinze and Schemper (2002).
}
\usage{
data_endometrial
}
\description{
A classic endometrial cancer dataset used to demonstrate separation in
logistic regression. The outcome is high histology grade. Neovascularization
is completely absent among low-grade cases in this dataset, making it useful
for teaching Firth penalized logistic regression.
}
\references{
Agresti A (2015). \emph{Foundations of Linear and Generalized Linear Models}.
Wiley.

Heinze G, Schemper M (2002). A solution to the problem of separation in
logistic regression. \emph{Statistics in Medicine}, 21, 2409-2419.
\doi{10.1002/sim.1047}
}
\keyword{datasets}
