% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data_lungcancer}
\alias{data_lungcancer}
\title{Lung Cancer Trial Data}
\format{
A data frame with 137 observations and 8 variables:
\describe{
  \item{trt}{Treatment group (1 = standard, 2 = test)}
  \item{celltype}{Cell type (squamous, smallcell, adeno, large)}
  \item{time}{Survival time (in days)}
  \item{status}{Censoring status (1 = died, 0 = censored)}
  \item{karno}{Karnofsky performance score (higher = better)}
  \item{diagtime}{Months from diagnosis to randomization}
  \item{age}{Age in years}
  \item{prior}{Prior therapy (0 = no, 10 = yes)}
}
}
\source{
\url{https://CRAN.R-project.org/package=survival}
}
\usage{
data_lungcancer
}
\description{
Survival data from a clinical trial of lung cancer patients conducted
 by the Veteran's Administration.
}
\references{
Kalbfleisch JD and Prentice RL (1980).
 The Statistical Analysis of Failure Time Data.
}
\keyword{datasets}
