% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders-flex.R
\name{.build_flex_strata_wide_multi}
\alias{.build_flex_strata_wide_multi}
\title{Flextable builder for stratified MULTIVARIABLE wide tables (2 cols/stratum)}
\usage{
.build_flex_strata_wide_multi(df, spanners, effect_label_adj, theme, footnotes)
}
\description{
Flextable builder for stratified MULTIVARIABLE wide tables (2 cols/stratum)
}
\keyword{internal}
