% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders-flex.R
\name{.build_flex_strata_wide_uni}
\alias{.build_flex_strata_wide_uni}
\title{Flextable builder for stratified wide tables (3 cols per stratum)}
\usage{
.build_flex_strata_wide_uni(df, spanners, effect_label, theme, footnotes)
}
\description{
Flextable builder for stratified wide tables (3 cols per stratum)
}
\keyword{internal}
