% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders-gt.R
\name{.build_gt_strata_wide_uni}
\alias{.build_gt_strata_wide_uni}
\title{GT builder for stratified wide tables (3 cols per stratum)}
\usage{
.build_gt_strata_wide_uni(df, spanners, effect_label, theme, footnotes)
}
\description{
GT builder for stratified wide tables (3 cols per stratum)
}
\keyword{internal}
