% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multi_model_nbin.R
\name{.fit_multi_model_nbin}
\alias{.fit_multi_model_nbin}
\title{Fit Multivariable Negative Binomial Model (Internal)}
\usage{
.fit_multi_model_nbin(data, outcome, exposures)
}
\arguments{
\item{data}{A `data.frame` with complete observations for outcome and exposures.}

\item{outcome}{A string. Name of the count outcome variable.}

\item{exposures}{A character vector of predictor (exposure) variable names.}
}
\value{
A fitted model object (of class `negbin`) or `NULL` if model fitting fails.
}
\description{
Fits a negative binomial regression using `glm.nb()` for a count outcome and multiple predictors.
}
\keyword{internal}
