% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_uni_model_nbin.R
\name{.fit_uni_model_nbin}
\alias{.fit_uni_model_nbin}
\title{Fit Univariate Negative Binomial Regression Model (Internal)}
\usage{
.fit_uni_model_nbin(data, outcome, exposure)
}
\arguments{
\item{data}{A `data.frame` with complete observations for outcome and exposures.}

\item{outcome}{A string. Name of the count outcome variable.}

\item{exposure}{A single predictor variable name.}
}
\value{
A fitted model object (of class `negbin`) or `NULL` if model fitting fails or data is insufficient.
}
\description{
Fits a univariate negative binomial regression model using `glm.nb()` for a count outcome and single predictor.
}
\keyword{internal}
