% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnotes.R
\name{.footnotes_uni_strata}
\alias{.footnotes_uni_strata}
\title{Compose footnotes for univariate strata}
\usage{
.footnotes_uni_strata(approach)
}
\description{
Compose footnotes for univariate strata
}
\keyword{internal}
