% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-strata.R
\name{.strata_build_wide_multi}
\alias{.strata_build_wide_multi}
\title{Assemble wide display DF for all strata (multivariable)}
\usage{
.strata_build_wide_multi(data, exposures, stratifier, td_by_stratum)
}
\description{
Assemble wide display DF for all strata (multivariable)
}
\keyword{internal}
