% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-strata.R
\name{.strata_pull_cols_multi}
\alias{.strata_pull_cols_multi}
\title{Pull (Adjusted Effect, p) for a multivariable stratum, aligned to skeleton}
\usage{
.strata_pull_cols_multi(td, skeleton, is_factor)
}
\description{
Pull (Adjusted Effect, p) for a multivariable stratum, aligned to skeleton
}
\keyword{internal}
