% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_reg_nbin.R
\name{multi_reg_nbin}
\alias{multi_reg_nbin}
\title{Multivariable Negative Binomial Regression (Adjusted IRRs)}
\usage{
multi_reg_nbin(data, outcome, exposures)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{outcome}{The name of the count outcome variable.}

\item{exposures}{A vector of predictor variables.}
}
\value{
A `gtsummary::tbl_regression` object with exponentiated estimates labeled as Adjusted IRRs,
along with `$models`, `$model_summaries`, and `$table`.
}
\description{
Calculates adjusted incidence rate ratios (IRRs) using Negative Binomial Regression.
}
