% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimaIndiansDiabetes2.R
\docType{data}
\name{PimaIndiansDiabetes2}
\alias{PimaIndiansDiabetes2}
\title{PimaIndians2 Diabetes Dataset}
\format{
A data frame with 768 observations and 9 variables:
\describe{
  \item{pregnant}{Number of times pregnant}
  \item{glucose}{Plasma glucose concentration (glucose tolerance test)}
  \item{pressure}{Diastolic blood pressure (mm Hg)}
  \item{triceps}{Triceps skin fold thickness (mm)}
  \item{insulin}{2-Hour serum insulin (mu U/ml)}
  \item{mass}{Body mass index (BMI)}
  \item{pedigree}{Diabetes pedigree function}
  \item{age}{Age in years}
  \item{diabetes}{Factor indicating diabetes status (pos/neg)}
}
}
\source{
\url{https://www.openml.org/d/37}
}
\usage{
PimaIndiansDiabetes2
}
\description{
A cleaned version of the original Pima Indians Diabetes dataset from the `mlbench` package.
Useful for demonstrating regression approaches for binary outcomes.
}
\keyword{datasets}
