% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.identify_confounder_result}
\alias{print.identify_confounder_result}
\title{Print identify_confounder results}
\usage{
\method{print}{identify_confounder_result}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{identify_confounder()}.}

\item{...}{Ignored. Present for compatibility with the generic.}
}
\description{
Print identify_confounder results
}
