% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Check Convergence for a Specific Regression Approach}
\usage{
check_convergence(
  data,
  exposures,
  outcome,
  approach = "logit",
  multivariate = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{exposures}{A character vector of predictor variables.}

\item{outcome}{The name of the outcome variable.}

\item{approach}{Regression approach: one of "logit", "log-binomial", "poisson",
"robpoisson" or "negbin".}

\item{multivariate}{Logical. If TRUE, checks convergence for a multivariate model.}
}
\value{
A data frame with columns: Exposure, Model, Converged, Max.prob.
}
\description{
This function checks convergence status and maximum predicted probability
for each exposure (univariate) or combined exposures (multivariate)
using a specified regression approach.
}
\details{
For `robpoisson`, predicted probabilities may occasionally exceed 1. These are valid for estimating risk ratios but should not be interpreted as actual probabilities.
}
