% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-strata.R
\name{.strata_build_wide_uni}
\alias{.strata_build_wide_uni}
\title{Assemble wide display DF for all strata (univariate)}
\usage{
.strata_build_wide_uni(data, outcome, exposures, stratifier, per_stratum)
}
\description{
Assemble wide display DF for all strata (univariate)
}
\keyword{internal}
