% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_confounder.R
\name{identify_confounder}
\alias{identify_confounder}
\title{Identify Confounder Using Crude and Stratified Models}
\usage{
identify_confounder(
  data,
  outcome,
  exposure,
  potential_confounder,
  approach = "logit",
  threshold = 10,
  method = c("change", "mh")
)
}
\arguments{
\item{data}{A data frame.}

\item{outcome}{The name of the outcome variable (binary, count, or continuous depending on approach).}

\item{exposure}{The name of the exposure variable (must be binary or categorical).}

\item{potential_confounder}{A single variable to stratify by for confounder assessment.}

\item{approach}{Regression approach: one of "logit", "log-binomial", "poisson", "negbin", "linear", "robpoisson", "margstd_boot", or "margstd_delta".}

\item{threshold}{Numeric value specifying the percentage change in effect size considered meaningful for confounding (default is 10).}

\item{method}{Method for confounder identification: "mh" for Mantel-Haenszel, "change" for change-in-estimate (default).}
}
\value{
A list (invisible) containing crude and adjusted or stratified estimates, percent change,
logical indicators for confounding and effect modification, and reason for NA values if applicable.
}
\description{
This function estimates the crude (unadjusted) effect of an exposure on a binary, count, or continuous outcome
and compares it with Mantel-Haenszel estimates from stratified models using a potential confounder,
or with adjusted models using the change-in-estimate method.
It evaluates confounding based on the percent change in effect size and provides a heuristic check for effect modification.
Supports multiple regression approaches including logit, log-binomial, Poisson, negative binomial, robust Poisson, and marginal standardization methods.
}
