% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_models.R
\name{interaction_models}
\alias{interaction_models}
\title{Compare models with and without interaction term}
\usage{
interaction_models(
  data,
  outcome,
  exposure,
  covariates = NULL,
  effect_modifier,
  approach = "logit",
  test = c("LRT", "Wald"),
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{outcome}{Outcome variable (binary, count, or continuous).}

\item{exposure}{Main exposure variable (must be length 1).}

\item{covariates}{Optional vector of adjustment variables (default = NULL).}

\item{effect_modifier}{Variable to test for interaction with the exposure.}

\item{approach}{Modeling approach: "logit", "log-binomial", "poisson", "robpoisson", "negbin", "linear".}

\item{test}{Test for model comparison: "LRT" (default) or "Wald".}

\item{verbose}{Logical, whether to print interpretation.}
}
\value{
A list of model objects, p-value, and interpretation.
}
\description{
This function compares two models: one with and one without an interaction term to evaluate whether the interaction between exposure and a potential effect modifier
significantly improves model fit.
}
