% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_reg_nbin.R
\name{multi_reg_nbin}
\alias{multi_reg_nbin}
\title{Multivariable Negative Binomial Regression (Adjusted Risk Ratios)}
\usage{
multi_reg_nbin(data, outcome, exposures, summary = FALSE)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{outcome}{The name of the outcome variable (count).}

\item{exposures}{A vector of predictor variables.}

\item{summary}{Logical; if `TRUE`, prints model summary. Default is `FALSE`.}
}
\value{
A publication-ready `gtsummary` table with "Adjusted IRR".
}
\description{
This function calculates adjusted risk ratios for multiple exposures using **Negative Binomial Regression**.
}
