% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg.R
\name{plot_reg}
\alias{plot_reg}
\title{Visualize a Regression Model (Forest Plot)}
\usage{
plot_reg(
  tbl,
  title = NULL,
  ref_line = 1,
  order_y = NULL,
  log_x = FALSE,
  xlim = NULL,
  breaks = NULL,
  point_color = "#1F77B4",
  errorbar_color = "#4C4C4C",
  base_size = 14,
  show_ref = TRUE
)
}
\arguments{
\item{tbl}{A gtsummary object (from uni_reg, multi_reg, etc.).}

\item{title}{Optional title.}

\item{ref_line}{Reference line (default = 1).}

\item{order_y}{Optional character vector for custom y-axis header order.}

\item{log_x}{Logical. Log-transform the x-axis (default = FALSE).}

\item{xlim}{Optional x-axis limits.}

\item{breaks}{Optional x-axis breaks.}

\item{point_color}{Point color.}

\item{errorbar_color}{Error bar color.}

\item{base_size}{Font size.}

\item{show_ref}{Logical. Whether to show reference categories. Default is TRUE.}
}
\value{
A ggplot2 object.
}
\description{
Handles both univariate and multivariate models with hierarchical labels.
}
