% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_combine.R
\name{plot_reg_combine}
\alias{plot_reg_combine}
\title{Visualize Univariate and Multivariate Regression Side-by-Side}
\usage{
plot_reg_combine(
  tbl_uni,
  tbl_multi,
  title_uni = "Unadjusted",
  title_multi = "Adjusted",
  ref_line = 1,
  xlim_uni = NULL,
  breaks_uni = NULL,
  xlim_multi = NULL,
  breaks_multi = NULL,
  point_color = "#1F77B4",
  errorbar_color = "#4C4C4C",
  base_size = 14,
  show_ref = TRUE
)
}
\arguments{
\item{tbl_uni}{A `gtsummary` object from `uni_reg()`.}

\item{tbl_multi}{A `gtsummary` object from `multi_reg()`.}

\item{title_uni}{Optional title for univariate plot.}

\item{title_multi}{Optional title for multivariate plot.}

\item{ref_line}{Reference line (default = 1).}

\item{xlim_uni, breaks_uni}{Axis limits and breaks for univariate plot.}

\item{xlim_multi, breaks_multi}{Axis limits and breaks for multivariate plot.}

\item{point_color}{Point color.}

\item{errorbar_color}{Error bar color.}

\item{base_size}{Font size.}

\item{show_ref}{Logical. Whether to show reference categories.}
}
\value{
A combined `ggplot` object.
}
\description{
Visualize Univariate and Multivariate Regression Side-by-Side
}
