% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_functions.R
\name{save_docx}
\alias{save_docx}
\title{Save Multiple Tables and Plots to a Word Document}
\usage{
save_docx(tables = NULL, plots = NULL, filename = "report.docx", titles = NULL)
}
\arguments{
\item{tables}{A list of gtsummary tables.}

\item{plots}{A list of ggplot2 plots.}

\item{filename}{File name (with or without `.docx` extension).}

\item{titles}{Optional. A character vector of titles before each table or plot.}
}
\value{
A Word document saved in the working directory.
}
\description{
Save multiple gtsummary tables and ggplot2 plots to a Word document.
}
\examples{
\dontrun{
tbl <- gtsummary::tbl_regression(glm(mpg ~ hp + wt, data = mtcars))
p <- ggplot2::ggplot(mtcars, ggplot2::aes(x = wt, y = mpg)) + ggplot2::geom_point()
save_docx(
  tables = list(tbl),
  plots = list(p),
  filename = "report.docx",
  titles = c("Table 1: Regression", "Figure 1: Scatterplot")
)
}
}
