% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_functions.R
\name{save_table}
\alias{save_table}
\title{Save a Single Regression Table}
\usage{
save_table(tbl, filename = "table", format = c("docx", "pdf", "html"))
}
\arguments{
\item{tbl}{A gtsummary table.}

\item{filename}{File name (with or without extension).}

\item{format}{One of `"docx"`, `"pdf"`, or `"html"`.}
}
\value{
A saved file in the working directory.
}
\description{
Save a gtsummary table as Word, PDF, or HTML file.
}
\examples{
\dontrun{
tbl <- gtsummary::tbl_regression(glm(mpg ~ hp + wt, data = mtcars))
save_table(tbl, filename = "regression_table", format = "docx")
}
}
