% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_multi_nbin.R
\name{stratified_multi_nbin}
\alias{stratified_multi_nbin}
\title{Stratified Multivariable Negative Binomial Regression}
\usage{
stratified_multi_nbin(data, outcome, exposures, stratifier)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{outcome}{A character string specifying the count outcome variable.}

\item{exposures}{A character vector of predictor (exposure) variables.}

\item{stratifier}{A character string specifying the stratifier}
}
\value{
An object of class `stratified_multi_reg_nbin` with the following:
\describe{
  \item{\code{$table}}{A `gtsummary::tbl_merge` object of stratified tables.}
  \item{\code{$models}}{A named list of `glm.nb` model objects per stratum.}
  \item{\code{$model_summaries}}{A list of tidy model summaries.}
}
}
\description{
Performs multivariable negative binomial regression
stratified by a specified variable.
Observations with missing values in the stratifier are excluded.
}
\section{Accessors}{

Use `object$table`, `object$models`, `object$model_summaries`,
or `object$reg_check` to extract components.
}

\seealso{
[multi_reg_nbin()], [stratified_uni_reg_nbin()]
}
