% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_multi_nbin.R
\name{stratified_multi_nbin}
\alias{stratified_multi_nbin}
\title{Stratified Multivariable Negative Binomial Regression}
\usage{
stratified_multi_nbin(data, outcome, exposures, stratifier)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{outcome}{The outcome variable (count).}

\item{exposures}{A character vector of exposure variables.}

\item{stratifier}{A categorical variable to stratify the data by.}
}
\value{
A `gtsummary::tbl_merge` object with attributes for models, summaries, and diagnostics.
}
\description{
Performs multivariable negative binomial regression of exposures on a count outcome,
stratified by a specified variable. NAs in the stratifier are excluded.
}
