% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_uni_nbin.R
\name{stratified_uni_nbin}
\alias{stratified_uni_nbin}
\title{Stratified Univariate Negative Binomial Regression}
\usage{
stratified_uni_nbin(data, outcome, exposures, stratifier)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{outcome}{The name of the count outcome variable.}

\item{exposures}{A character vector of predictor variables.}

\item{stratifier}{A categorical variable to stratify the data by.}
}
\value{
A `gtsummary::tbl_merge` object with components: `$table`, `$models`, `$model_summaries`,
}
\description{
Performs univariate negative binomial regression of exposures on outcome,
stratified by a specified variable. NAs in the stratifier are excluded.
}
