% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_uni_nbin.R
\name{stratified_uni_nbin}
\alias{stratified_uni_nbin}
\title{Stratified Univariate Negative Binomial Regression}
\usage{
stratified_uni_nbin(data, outcome, exposures, stratifier)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{outcome}{A character string specifying the count outcome variable.}

\item{exposures}{A character vector specifying the exposure variables.}

\item{stratifier}{A character string specifying the stratifying variable.}
}
\value{
An object of class `stratified_uni_reg_nbin` with the following:
\describe{
  \item{\code{$table}}{A `gtsummary::tbl_merge` tables by stratum.}
  \item{\code{$models}}{A named list of fitted negbin models per stratum.}
  \item{\code{$model_summaries}}{A list of tidy summaries for each model.}
}
}
\description{
Performs univariate negative binomial regression
stratified by a specified variable.
Missing values in the stratifier are excluded.
}
\section{Accessors}{

Use `object$table`, `object$models`, or
`object$model_summaries` to extract components.
}

\examples{
if (requireNamespace("MASS", quietly = TRUE)) {
  data(Pima.tr, package = "MASS")
  Pima.tr$strata <- ifelse(Pima.tr$bmi > 30, "High", "Normal")
  strat_uni_nb <- stratified_uni_nbin(
    data = Pima.tr,
    outcome = "glu",
    exposures = c("age", "ped"),
    stratifier = "strata"
  )
  strat_uni_nb$table
}
}
\seealso{
[uni_reg_nbin()], [stratified_multi_reg()], [check_dispersion()]
}
