% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_uni_reg.R
\name{stratified_uni_reg}
\alias{stratified_uni_reg}
\title{Stratified Univariate Regression (Odds, Risk, or Rate Ratios)}
\usage{
stratified_uni_reg(
  data,
  outcome,
  exposures,
  stratifier,
  approach = "logit",
  summary = FALSE
)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{outcome}{The name of the outcome variable.}

\item{exposures}{A character vector of predictor variables.}

\item{stratifier}{The name of the variable to stratify results by.}

\item{approach}{Modeling approach to use. One of:
`"logit"`, `"log-binomial"`, `"poisson"`, `"robpoisson"`, `"linear"`}

\item{summary}{Logical; if `TRUE`, prints model summaries for each stratum. Default is `FALSE`.}
}
\value{
A `gtsummary::tbl_merge` object with unadjusted estimates per stratum.
}
\description{
Performs univariate regression for each exposure on a binary, count, or continuous outcome,
stratified by a specified variable. Produces one column per stratum.
}
