% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_reg.R
\name{uni_reg}
\alias{uni_reg}
\title{Univariate regression (Odds, Risk, or Rate Ratios)}
\usage{
uni_reg(data, outcome, exposures, approach = "logit", summary = FALSE)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{outcome}{The name of the outcome variable (binary or count).}

\item{exposures}{A vector of predictor variables.}

\item{approach}{Modeling approach to use. One of:
`"logit"` (OR), `"log-binomial"` (RR), `"poisson"` (IRR),
`"robpoisson"` (RR), `"linear"` (Beta coefficients)}

\item{summary}{Logical; if `TRUE`, prints model summaries for each univariate model. Default is `FALSE`.}
}
\value{
A `gtsummary::tbl_stack` object of exponentiated unadjusted estimates.
}
\description{
Performs univariate regression for each exposure on a binary or count outcome.
Depending on `approach`, returns either Odds Ratios (OR), Risk Ratios (RR), or Incidence Rate Ratios (IRR).
}
\details{
This function requires the following packages: `dplyr`, `purrr`, `gtsummary`, `risks`.
}
