% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_reg_nbin.R
\name{uni_reg_nbin}
\alias{uni_reg_nbin}
\title{Univariate Negative Binomial Regression}
\usage{
uni_reg_nbin(data, outcome, exposures, summary = FALSE)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{outcome}{The name of the count outcome variable.}

\item{exposures}{A vector of predictor variables.}

\item{summary}{Logical; if `TRUE`, prints model summaries. Default is `FALSE`.}
}
\value{
A stacked summary table of negative binomial models.
}
\description{
This function fits negative binomial regression models for multiple exposures.
It returns a publication-ready table using `gtsummary`.
}
